Kotlin study - DataBase

+++ Kotlin/AluGames/src/main/kotlin/br/com/alura/alugames/modelo/InfoJogoJson.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 data class InfoJogoJson(
     val titulo: String,
     val capa: String,
-    val preco: BigDecimal,
+    val preco: Double,
     val descricao: String)
